Update PhysFormerTrainer.py (#404)
* Update PhysFormerTrainer.py

Bug in "train" function in PhysFormerTrainer. The "get_hr" function is used in line 105 to get the ground truth Heart Rate using the ground truth ppg signal. This function uses the "sr" variable for sampling rate. This should be set according to the frame rate of the ground truth ppg. Currently, it is set to 30 and this is used for all datasets leading to incorrect heart rates. 
Modified line 105 to use the correct sampling or frame rate each time:
"get_hr(i, sr=self.frame_rate)"

* Update PhysFormerTrainer.py

Hi,

I created a new variable to get the validation frame_rate in case someone uses different datasets for validation and training.
        self.frame_rate_valid = config.VALID.DATA.FS


I used this value inside the validation function:
                    hrs.append((self.get_hr(_1.cpu().detach().numpy(), sr=self.frame_rate_valid), self.get_hr(_2.cpu().detach().numpy(), sr=self.frame_rate_valid)))